+++ frmOptionsForm.pas
@@ -40,7 +40,6 @@ type
     cbTrackAllEditorID: TCheckBox;
     cbSortINFO: TCheckBox;
     cbShowFlagEnumValue: TCheckBox;
-    cbRemoveOffsetData: TCheckBox;
     pnlFontRecords: TPanel;
     pnlFontMessages: TPanel;
     pnlFontViewer: TPanel;